+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Profiles can be reconfigured via the `--reconf` command line argument.
+
 ## 2.2.0 - 2022-10-15
 
 ### Features